+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Reorganize how routing related features are represented internally to simplify features management.
 - Fix a case where routing tags would not be injected when given routing definition would not be used with a block
 - Fix a case where using `#active_job_topic` without extra block options would cause `manual_offset_management` to stay false.
+- Fix a case when upon Pro ActiveJob usage with Virtual Partitions, correct offset would not be stored
 
 ## 2.0.15 (2022-10-20)
 - Sanitize admin config prior to any admin action.